Added updated version of the project file
Updated code for displaying the image at the end of the program. Single image path chosen. Imshow needs a tensor passed to it in order to work. 

Still figuring out a way to select image path manually and pass it to the imshow function as a tensor after the network has been trained.